some work on removing stl.h

+++ src/emitters/sun.cpp
@@ -178,7 +178,7 @@ public:
 		Spectrum *target = (Spectrum *) bitmap->getFloatData();
 		Spectrum value =
 			m_radiance * (2 * M_PI * (1-std::cos(m_theta))) *
-			(bitmap->getWidth() * bitmap->getHeight())
+			(Float) (bitmap->getWidth() * bitmap->getHeight())
 			/ (2.0f * M_PI * M_PI * (Float) nSamples);
 
 		for (size_t i=0; i<nSamples; ++i) {
